requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch
Pillow
numpy
matplotlib
scikit-learn
tqdm
